Extract out readers of binary data into submodule
