+++ corpus_rust/src/corpus_query_engine.rs
@@ -1,4 +1,7 @@
-use crate::common::{PackedBitMask, u32_from_bytes, u64_from_bytes};
+mod corpus_data_readers;
+
+use crate::common::PackedBitMask;
+use crate::corpus_query_engine::corpus_data_readers::{CorpusText, IndexBuffers, TokenStarts};
 use crate::corpus_serialization::StoredMapValue;
 use crate::packed_index_utils::{apply_or_to_indices, smear_bitmask};
 use crate::query_parsing_v2::{
@@ -10,8 +13,6 @@ use super::common::IndexData;
 use super::corpus_serialization::LatinCorpusIndex;
 use super::packed_index_utils::apply_and_to_indices;
 use super::profiler::TimeProfiler;
-
-use crate::byte_readers::{InMemoryReader, MmapReader, RawByteReader};
 
 use serde::Serialize;
 use std::cmp::{max, min};
@@ -97,94 +98,10 @@ struct IntermediateResult {
     position: u32,
 }
 
-struct TokenStarts {
-    reader: MmapReader,
-}
-
-impl TokenStarts {
-    pub fn new(token_starts_path: &str) -> Result<Self, Box<dyn Error>> {
-        let reader = MmapReader::new(token_starts_path)?;
-        Ok(TokenStarts { reader })
-    }
-
-    pub fn token_start(&self, token_id: u32) -> Result<usize, String> {
-        let i = ((token_id * 2) * 4) as usize;
-        Ok(u32_from_bytes(self.reader.bytes(i, i + 4))?[0] as usize)
-    }
-
-    pub fn break_start(&self, token_id: u32) -> Result<usize, String> {
-        let i = ((token_id * 2 + 1) * 4) as usize;
-        Ok(u32_from_bytes(self.reader.bytes(i, i + 4))?[0] as usize)
-    }
-}
-
-struct CorpusText {
-    reader: MmapReader,
-}
-
-impl CorpusText {
-    pub fn new(raw_text_path: &str) -> Result<Self, Box<dyn Error>> {
-        let reader = MmapReader::new(raw_text_path)?;
-        Ok(CorpusText { reader })
-    }
-
-    pub fn slice(&self, start: usize, end: usize) -> String {
-        // We store the starts of each word as byte offsets, so this should always be valid UTF-8.
-        unsafe { String::from_utf8_unchecked(self.reader.bytes(start, end).to_vec()) }
-    }
-
-    pub fn advise_range(&self, start: usize, end: usize) {
-        self.reader.advise_range(start, end);
-    }
-}
-
-struct RawBuffers {
-    reader: InMemoryReader,
-}
-
-impl RawBuffers {
-    pub fn new(raw_buffer_path: &str) -> Result<Self, Box<dyn Error>> {
-        let reader = InMemoryReader::new(raw_buffer_path)?;
-        Ok(RawBuffers { reader })
-    }
-
-    pub fn resolve_index(
-        &self,
-        data: &StoredMapValue,
-        num_tokens: u32,
-    ) -> Result<IndexData, String> {
-        match data {
-            StoredMapValue::Packed { offset, len } => Ok(IndexData::Unpacked(
-                u32_from_bytes(
-                    self.reader
-                        .bytes(*offset as usize, (*offset + (4 * *len)) as usize),
-                )?
-                .to_vec(),
-            )),
-            StoredMapValue::BitMask { offset, .. } => {
-                let num_words = (num_tokens as usize).div_ceil(64);
-                let bytes = self
-                    .reader
-                    .bytes(*offset as usize, *offset as usize + (num_words * 8));
-                Ok(IndexData::PackedBitMask(PackedBitMask {
-                    data: u64_from_bytes(bytes)?.to_vec(),
-                }))
-            }
-        }
-    }
-
-    pub fn num_elements(&self, data: &StoredMapValue) -> u32 {
-        match data {
-            StoredMapValue::Packed { len, .. } => *len,
-            StoredMapValue::BitMask { num_set, .. } => *num_set,
-        }
-    }
-}
-
 pub struct CorpusQueryEngine {
     corpus: LatinCorpusIndex,
     text: CorpusText,
-    raw_buffers: RawBuffers,
+    raw_buffers: IndexBuffers,
     starts: TokenStarts,
 }
 
@@ -403,7 +320,7 @@ impl CorpusQueryEngine {
 impl CorpusQueryEngine {
     pub fn new(corpus: LatinCorpusIndex) -> Result<Self, Box<dyn Error>> {
         let text = CorpusText::new(&corpus.raw_text_path)?;
-        let raw_buffers = RawBuffers::new(&corpus.raw_buffer_path)?;
+        let raw_buffers = IndexBuffers::new(&corpus.raw_buffer_path)?;
         let starts = TokenStarts::new(&corpus.token_starts_path)?;
         Ok(CorpusQueryEngine {
             corpus,
